Add const/restrict, fix QWORD writes, optimize out strcmp() calls
Add const and restrict keywords in appropriate places.

Quad-word (64-bit) data writes were succeeding but throwing write
errors because the error check was inverted. QWORD write status
now correctly reflects the result of a write.

The strcmp() call was only being used for one purpose: checking
to see if a path consisted of a slash and nothing else. Due to the
simplicity of the check, it is better to look for a slash and a
null byte instead of incurring the call overhead. A macro was
defined to do the check and strcmp() calls replaced with it.