Improve flushing code, silence unused variable warnings
This adds fsync() and makes it flush dirty pages if called. It
also extends debug logging to include information about calls to
flush_dirty_pages() so we can see when that is happening.

A lot of parameters go unused and generate warnings. These are now
voided to silence compiler warnings about them.